app initialization and style
