+++ Screens/Auth/Login.js
@@ -5,8 +5,8 @@ import { initializeApp } from "firebase/app";
 import { useState } from "react";
 import { useNavigation } from "@react-navigation/native";
 import { StylesAuth } from "./styles";
+import app from "../../db/firebaseConfig";
 
-const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
 
 export function Login() {
