+++ db/firebaseConfig.js
@@ -14,4 +14,6 @@ export const firebaseConfig = {
 };
 
 // Initialize Firebase
-//const app = initializeApp(firebaseConfig);
+const app = initializeApp(firebaseConfig);
+
+export default app;
