+++ package.json
@@ -12,15 +12,16 @@
     "@react-navigation/bottom-tabs": "^6.4.0",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
-    "native-base": "^3.4.13",
     "expo": "~46.0.16",
     "expo-status-bar": "~1.4.0",
     "firebase": "^9.13.0",
+    "native-base": "^3.4.13",
     "react": "18.0.0",
     "react-native": "0.69.6",
     "react-native-gesture-handler": "^2.8.0",
-    "react-native-safe-area-context": "^4.4.1",
-    "react-native-screens": "^3.18.2"
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-screens": "^3.18.2",
+    "react-native-svg": "12.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
