+++ styles.js
@@ -1,34 +1,75 @@
 import { StyleSheet } from "react-native";
 
 export const Styles = StyleSheet.create({
-    Text: {
-        fontSize: 25,
-        fontWeight: "bold"
-    },
-    containerClients: {
-      flex: 1,
-      backgroundColor: '#A9A9A9',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    containerHome: {
-        flex: 1,
-        backgroundColor: '#FFE4E1',
-        alignItems: 'center',
-        justifyContent: 'center'
-    },
-    containerProducts: {
-        flex: 1,
-        backgroundColor: '#87CEFA',
-        alignItems: 'center',
-        justifyContent: 'center'
-    },
-    containerSales: {
-        flex: 1,
-        backgroundColor: '#008B8B',
-        alignItems: 'center',
-        justifyContent: 'center'
-    }
-  });
+  Text: {
+    fontSize: 25,
+    fontWeight: "bold",
+  },
+  containerClients: {
+    flex: 1,
+    backgroundColor: "#A9A9A9",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  containerHome: {
+    flex: 1,
+    backgroundColor: "#FFE4E1",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  containerProducts: {
+    flex: 1,
+    backgroundColor: "#87CEFA",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  containerSales: {
+    flex: 1,
+    backgroundColor: "#008B8B",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+  },
+  Title: {
+    marginTop: 50,
+    fontSize: 30,
+    fontWeight: "bold",
+  },
+  Middle: {
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  form: {
+    marginLeft: 10,
+    marginRight: 10,
+  },
+  buttonStyle: {
+    marginTop: 12,
+    marginLeft: 15,
+    marginRight: 15,
+  },
+  buttonDesign: {
+    backgroundColor: "red",
+  },
+  buttonText: {
+    color: "white",
+    fontWeight: "bold",
+  },
+  imageStyle: {
+    width: 80,
+    height: 80,
+    marginLeft: 20,
+  },
+  boxStyle: {
+    flexDirection: "row",
+    marginTop: 30,
+    marginLeft: 15,
+    marginRight: 15,
+    justifyContent: "space-around",
+  },
+});
 
-  //#00FA9A
+//#00FA9A
